+++ src/screens/send-flow/ReviewTransaction.tsx
@@ -347,8 +347,9 @@ export const ReviewTransaction: React.FC = () => {
           recallBlocks ? parseInt(recallBlocks) : undefined,
           isDelegateProofEnabled()
         ] as const;
-        const txId =
-          authorization === undefined
+        const txId = verifiedName
+          ? await initiateSendTransaction(...commonArguments, authorization, verifiedName)
+          : authorization === undefined
             ? await initiateSendTransaction(...commonArguments)
             : await initiateSendTransaction(...commonArguments, authorization);
         if (isExtension()) requestSWTransactionProcessing();
@@ -387,6 +388,7 @@ export const ReviewTransaction: React.FC = () => {
       publicKey,
       recallBlocks,
       sharePrivately,
+      verifiedName,
       to,
       token
     ]